refactor(events): address PR review feedback
Frontend:
- Split event-form.tsx (~950 lines) into event-form-schema.ts, a
  useEventForm hook, and event-form-sections.tsx, plus a shared FieldError
  component and clearLocation helper.
- Group the scheduled-event fields in a nested sub-box (still gated by the
  "Publicly listed" checkbox); stop force-collapsing details on save.
- Show timezone abbreviations in the picker options; drop the redundant
  "Times are in…" line. Remove the "Event details" eyebrow and assorted
  help/boilerplate text.
- Home hub: only render the events section for users with attendance access.
- Rename lib/timezone.ts -> lib/time.ts.
- Drop the frontend HH:MM:SS slice now that the backend returns HH:MM.
- Convert the mount-gate and places-autocomplete syncs off setState-in-effect
  to satisfy the react-hooks lint rules.

Backend:
- Return start_time/end_time as HH:MM from EventJSON.
- Use ValidationErrorf for public-event input validation (400 not 500).
- Tidy Google Places key comments.
