+++ frontend-v2/src/app/authed-page-provider.tsx
@@ -5,7 +5,6 @@ import { type AuthedUser } from '@/lib/api'
 
 export type TAuthedPageContext = {
   user: AuthedUser
-  // Referrer-restricted Google Places key, served from the backend config.
   googlePlacesApiKey: string
 }
